organization: move InletVector to its own file
